refactor: move v0_to_v1 migration to Converters folder
